a2aAgents/backend/requirements_integration.txt
# Integration dependencies for A2A Agents

# Core dependencies
aioetcd3>=1.10
pydantic>=2.0
pydantic-settings>=2.0
httpx>=0.24.0
asyncio>=3.11
web3>=6.0.0
eth-account>=0.8.0

# Storage and cache
aioredis>=2.0.0
aiocache>=0.12.0

# Message queue
aio-pika>=9.0.0
kombu>=5.3.0

# Monitoring and telemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-httpx>=0.41b0
prometheus-client>=0.18.0

# Security
cryptography>=41.0.0
PyJWT>=2.8.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
orjson>=3.9.0